Move privateroute to utils
