+++ src/utils/PrivateRoute.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
-import { useAuth } from "./SupaContext";
-const { session } = useAuth();
+import { useAuth } from "../context/SupaContext";
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
+  const { session } = useAuth();
   return (
     <Route
       {...rest}
